pyproject.toml
[project]
name = "ai-powered-resume-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.0",
    "flask-limiter>=3.12",
    "google-generativeai>=0.8.5",
    "python-docx>=1.1.2",
]
